Module cargo_auto_lib::auto_github_mod

Constants

GITHUB_TOKEN_PATH

RELEASES_MD

Functions

add_message_to_unreleased

add commit message to Unreleased in RELEASES.md

ask_for_identity_file_path_for_ssh

Ask the user for the filename of the ssh key used to connect with SSH/git to a server. host_name is like: github.com or bestia.dev, default like ~/.ssh/github_com_ssh_1 and ~/.ssh/bestia_dev_ssh_1 returns PathBuf to identity_file_path or None

body_text_from_releases_md

First, the user must write the content into file RELEASES.md in the section ## Unreleased. Then the automation task will copy the content to GitHub release and create a new Version title in RELEASES.md.

check_or_get_github_token 🔒

decrypt the token from GITHUB_TOKEN_PATH file or ask user interactively to type it, then encrypt and save to file

create_releases_md_if_file_not_exist 🔒

create RELEASES.md if file not exist

description_and_topics_to_github

Check and modify the description and topics on Github The words topics, keywords and tags all mean the same concept. In cargo.toml we have keywords. In README.md I want to have badges for tags In GitHub they are topics. Topic must be only one word: lowercase letters, hyphens(-) or numbers, less then 35 characters.

get_identity_file_path_from_ssh_config

Parse the ~/.ssh/config file and find the record for host_name and there is the identity_file_path. returns: identity_file_path contains the path like: ~/.ssh/github_com_ssh_1

git_tag_sync_check_create_push

sync, check, create, push git tag

github_api_create_new_release

create new release on Github
return release_id
https://docs.github.com/en/github/authenticating-to-github/keeping-your-account-and-data-secure/creating-a-personal-access-token

github_api_get_repository 🔒

github api get repository

github_api_replace_all_topics 🔒

github_api_repository_new

creates a new github repository

github_api_update_description

github api update description and topics

github_api_upload_asset_to_release

upload asset to github release
release_upload_url example: https://uploads.github.com/repos/owner/repo/releases/48127727/assets
https://docs.github.com/en/github/authenticating-to-github/keeping-your-account-and-data-secure/creating-a-personal-access-token
async function can be called from sync code:

init_repository_if_needed

init repository if needed. A new local git repository and remote GitHub repository.

new_local_repository

interactive ask to create a new local git repository

new_remote_github_repository

interactive ask to create a new remote GitHub repository

now_utc_date_iso

the UTC date in iso standard 2024-12-31

ssh_add_if_needed

identity_private_file_path contains the path of the private key like: ~/.ssh/github_com_ssh_1 check if this file is in ssh-add. if is not, then ask user to ssh-add and enter passcode.

ssh_add_resolve

Parse the ~/.ssh/config file and finds the record for host_name and there is the identity_file_path. If not found, ask user for identity_file_path, Check if this identity is already in ssh-agent and if not then run ssh-add to add it. This ssh-add will stay even after the process ends, so the parent process will still have it. returns: fingerprint or None and identity_file_name