Module cargo_auto_lib::auto_lines_of_code_mod

inserts shield badges with lines_of_code into README.rs

Structs

CargoToml

LinesOfCode
Struct that contains 4 types of lines count: code, doc comments, comments, test and examples.

Workspace

Functions

auto_lines_of_code
inserts shield badges with lines_of_code into README.rs the parameter Link to include in shield badge. If empty_string, the git remote repository will be used. Lines of code are not a “perfect” measurement of anything.
Anybody can write a very big number of lines of useless code and comments.
But for 95% of the cases they are good enough.
Most of the developers use some “standard” coding practices and that is quantifiable and comparable.

git_remote_output

include_into_readme_md
Includes (writes, modifies) the shield badge code into README.md file.

one_project_count_lines
private function. Use public workspace_or_project_count_lines().

process_git_remote
Return the string for link for badges like: https://github.com/LucianoBestia/auto_lines_of_code/.
Get the output string after $ git remote -v.
Then finds out the link to the repository with regex.
Returns empty string if something goes wrong: no git, no remote,…

regex_capture
returns a Result. in the case of error the calling fn will return empty string.

text_to_include

to_string_as_shield_badges
Returns a string with the markdown code for 4 shield badges.

workspace_or_project_count_lines
Returns the struct LinesOfCode for 4 types of lines: code, doc comments, comments, test and examples. Automatically detects if this is a workspace or single rust project.