cargo_auto_lib

Module auto_md_to_doc_comments_mod

Finds rs files with markers and include segments from md files as doc comments.

Structs

MdSegment 🔒

Markers found in md files

REGEX_MD_END 🔒

Regex for end marker

REGEX_MD_START 🔒

Regex for start marker

REGEX_RS_END 🔒

Regex for end marker

REGEX_RS_START 🔒

Regex for start marker

RsMarker 🔒

Markers found in rs files

Functions

auto_md_to_doc_comments

Finds rs files with markers and include segments from md files as doc comments.

get_md_segments_using_cache 🔒

The first time it is called read the file and extracts all the segments into a cache vector.

rs_file_markers 🔒

markers in rs files

rs_files 🔒

All rs files in src, tests and examples folders.