Function cargo_auto_lib::auto_cargo_toml_to_md

pub fn auto_cargo_toml_to_md()

includes data from Cargo.toml to README.md files: version, authors,… It works for workspaces and for single projects.
To avoid out of sync data like version, authors and description in the README.md files, auto_cargo_toml_to_md includes this data from Cargo.toml.
Run it on every build with cargo auto.
It works also with other md files in the project, not only README.md.
In the md file write these markers in markdown comments (invisible),
don’t copy the numbers 1 and 2:

1 [//]: # (auto_cargo_toml_to_md start)
2 [//]: # (auto_cargo_toml_to_md end)

auto_cargo_toml_to_md deletes the old lines between the markers and includes the Cargo.toml data:
description, repository, version, &utc_now(), authors

Run the example:

cargo run --example example_01_auto_cargo_toml_to_md