cargo_auto_lib

Module utils_mod

Functions for various utilities.

Functions

concatenate_vec_to_string

The original concat() function does not have a delimiter.

find_from

Find from pos

find_pos_end_data_before_delimiter

Position of end of the delimited data before the delimiter

find_pos_start_data_after_delimiter

Position of start of the delimited data after the delimiter

now_utc_date_iso

UTC date in iso standard like 2024-12-31.

traverse_dir_with_exclude_dir

Traverse dir and its sub-dir, but avoid excluded dirs.