Module cargo_auto_lib::auto_encrypt_decrypt_with_ssh_mod

How to save a GitHub TOKEN securely inside a file? GitHub TOKEN is used by GitHub API to gain access (authentication and authorization) to your GitHub. A TOKEN is a secret just like a password and it must be protected. If the TOKEN is leaked, a mal-intentioned can gain access to everything, using the API. Never store TOKENS in plain text anywhere! The TOKEN must be encrypted before storing it. We already use SSH keys to connect to GitHub. And we use ssh-agent for easy work with SSH identities. I want to use the ssh key inside ssh-agent to encrypt the TOKEN and save it in a file. The easy solution: encrypt with the Public key and then decrypt with the Private key. Problem: ssh-agent can only sign a message with the private key. Nothing more. It cannot decrypt with private key, because it would be a security risk. The security is based on the assumption that only the owner of the ssh private key can sign the message. The user already uses the ssh private key and it uses ssh-agent to connect over ssh to GitHub. So the user already knows how important are ssh private keys and to keep them secure. This could work also for other TOKENS, not just GitHub. Encryption

ssh-agent must contain the chosen ssh identity. Use ssh-add for this.

Create a random message that will be used only by this code. It is not a secret.

Sign this message. This will become the password for symmetric encryption. It is a secret.

Input the token interactively in hidden input. It is a secret.

Use the password to symmetric encrypt the token.

Zeroize the token and password.

Save the message and the encrypted token in a json file. Decryption

ssh-agent must contain the ssh identity. Use ssh-add for this.

Read the json file, get the ssh_identity_file_path, message and the encrypted token.

Find the right identity inside ssh-agent.

Sign the message. This will become the password for symmetric decryption. It is a secret.

Use this password to symmetric decrypt the token.

Get the token.

Zeroize the token and password.

Structs

EncryptedString

SecretBytes

SecretString

Functions

decrypt_with_ssh_from_json 🔒

decrypt from json file with ssh_identity use ssh-add to put ssh identity into ssh-agent

encrypt_with_ssh_interactive_save_json 🔒

encrypt a token with the chosen ssh_identity and save as json file use ssh-add to put ssh identity into ssh-agent

file_exists 🔒

check if file exists

file_path_home_expand 🔒

expands the ~ for home_dir and returns expanded path

ssh_add_contains_public_key 🔒

returns the public_key inside ssh-add or panics