Module cargo_auto_lib::auto_md_to_doc_comments_mod

finds rs files with markers and include segments from md files It works for workspaces and for single projects.

Structs

MdSegment 🔒

REGEX_MD_END 🔒

REGEX_MD_START 🔒

REGEX_RS_END 🔒

REGEX_RS_START 🔒

RsMarker 🔒

Functions

auto_md_to_doc_comments

finds rs files with markers and include segments from md files Includes segments of md files into rs files as doc comments.
It works with workspaces and single projects. From this doc comments cargo doc will generated the documentation and auto-completion.
We don’t want to manually copy this segments. We want them to be automatically in sync.
We will just run this function before every cargo doc with an automation task.
The auto_md_to_doc_comments function must be executed in the project root folder where is the Cargo.toml file.
First it searches all the rs files in src, tests and examples folders.
If they contain the markers, than finds the md file and the named segment and include it as doc comments into the rs file.
The markers are always in pairs: start and end. So exactly the content in between is changed. The markers are always comments, so they don’t change the code.
It works only for files with LF line delimiter. No CR and no CRLF.

get_md_segments_using_cache 🔒

The first time it is called reads the file and extracts all the segments into a cache vector. Subsequent calls read from the cache.

one_project 🔒

rs_file_markers 🔒

markers in rs files

rs_files 🔒

All rs files in src, tests and examples folders. The current dir must be the project root where the Cargo.toml is.