Module cargo_auto_lib::auto_version_from_date_mod

new version as date is written to Cargo.toml and service_worker.js It works for workspaces and for single projects.

Structs

AutoVersionFromDate

the struct that represents the file .automation_tasks_rs_file_hashes.json

FileMetaData

file metadata

Functions

are_files_equal

if files are added or deleted, other files must be also changed I need to check if the files on the filesystem are the same as in the json

auto_version_from_date

auto_version_from_date

auto_version_from_date_forced

Works for single projects and workspaces.
Just like auto_version_from_date(), but force the new version even if no files are changed. For workspaces release I want to have the same version in all members.
It is slower, but easier to understand when deployed.

auto_version_from_date_internal 🔒

correct_file_metadata_for_cargo_tom_inside_vec

the Cargo.toml is now different and needs to be changed in the vec of file metadata

do_one_project 🔒

find_from 🔒

in string find from position

modify_service_js 🔒

search for file service_worker.js and modify version

read_file_metadata

make a vector of file metadata

read_json_file

read .automation_tasks_rs_file_hashes.json

save_json_file_for_file_meta_data

save the new file metadata

sha256_digest 🔒

calculate the hash for a file

version_from_date 🔒

converts a date to a version

write_version_to_cargo_and_modify_metadata 🔒

move vec_of_metadata