cargo_auto_lib
Module auto_version_from_date_mod

The new version as date is written to Cargo.toml and service_worker.js.

Structs

AutoVersionFromDate

The struct that represents the file .automation_tasks_rs_file_hashes.json

FileMetaData

File metadata

Functions

are_files_equal

If files are added or deleted, other files must be also changed

auto_version_from_date

New version from date is written to Cargo.toml and service_worker.js.

auto_version_from_date_forced

Just like auto_version_from_date(), but force the new version even if no files are changed.

auto_version_from_date_internal 🔒

Internal function to get version from date

correct_file_metadata_for_cargo_tom_inside_vec

Cargo.toml is now different and needs to be changed in the vec of file metadata

find_from 🔒

Find from position in string

modify_service_js 🔒

Search for file service_worker.js and modify version

read_file_metadata

Make a vector of file metadata

read_json_file

Read .automation_tasks_rs_file_hashes.json

save_json_file_for_file_meta_data

Save the new file metadata

sha256_digest 🔒

Calculate the hash for the content of a file

version_from_date 🔒

Convert a date to a version

write_version_to_cargo_and_modify_metadata 🔒

Write version to Cargo.toml