Module cargo_auto_lib::public_api_mod

public_api_mod

The Public API of a library is a pain in the a…
Every time I modify something I have to think how it will affect the users of the library.
They could have tens or hundreds of places where they use the library. Breaking changes are terrible things.
The developers are not willing to change their code every time a library changes slightly.
Yes, there is the semver to show if the new library is API compatible. That helps a lot.
It is dreaded if the first version of a function does not return a Result<>.
Then later we will surely come to the point, that we need to return a Result<>. This is a terrible breaking change.
It is wise to return a Result always. Even when that is not needed right now. It will surely be needed in the future.
Another tactic is to make new functions with a different name and mark the old functions as Obsolete.

This library is used by the automation_tasks_rs executable.
I want to have here the complete and exact definition of the public API.
Therefore I will not use reexports like pub use or pub mod.
This way I can always know easily if my public API has changed.
Just compare the public_api_mod.rs file in git diff.
Adding functions, structs, methods and enums is ok, it does not break the Public API.
But modifying existing functions, methods or enums will break the compatibility.

Re-exports

pub use crate::error_mod::ResultWithLibError;

pub use crate::auto_helper_functions_mod::ShellOutput;

pub use crate::auto_ssh_mod::SecretString;

pub use crate::auto_cargo_toml_mod::CargoToml;

Constants

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Traits

CargoTomlPublicApiMethods

Trait with methods to read data from Cargo.toml

Functions

add_message_to_unreleased

Add commit message to Unreleased in RELEASES.md

auto_cargo_toml_to_md

This function includes data from Cargo.toml to markdown files.

auto_check_micro_xml

I want html pages to be correct microXML when I use them for single page application.

auto_delete_old_js_snippets

deletes old js snippets when working with wasm-pack

auto_doc_tidy_html

Pretty HTML for docs

auto_lines_of_code

This function inserts shield badges with lines_of_code into README.rs.

auto_md_to_doc_comments

Finds rs files with markers and include segments from md files as doc comments

auto_plantuml

process plantuml in current directory

auto_plantuml_for_path

Process plantuml for all md files

auto_playground_run_code

Include the link to run code in Rust playground

auto_semver_increment_minor

Increment the minor version in Cargo.toml file only if files are changed

auto_semver_increment_minor_forced

Increment the minor version in Cargo.toml file even if files are not changed

auto_semver_increment_patch

Increment the patch version in Cargo.toml file only if files are changed

auto_semver_increment_patch_forced

Increment the patch version in Cargo.toml file even if files are not changed

auto_version_from_date

New version from date is written to Cargo.toml and service_worker.js

auto_version_from_date_forced

Just like auto_version_from_date(), but force the new version even if no files are changed.

auto_version_increment_semver_or_date

Increment the version in Cargo.toml

auto_version_increment_semver_or_date_forced

Increment the version in Cargo.toml

body_text_from_releases_md

Get release text from RELEASES.md

completion_return_one_or_more_sub_commands

println one, more or all sub_commands

concatenate_vec_to_string

The original concat() function does not have a delimiter

copy_folder_files_into_module

Copy all files from the folder into a module as strings (static &str)

description_and_topics_to_github

Check and modify the description and topics on Github

exit_if_not_run_in_rust_project_root_directory

Check if the code was run inside the Rust project root directory.

find_from

find from_pos

find_pos_end_data_before_delimiter

return the position of end of the delimited data before the delimiter

find_pos_start_data_after_delimiter

return the position of start of the delimited data after the delimiter

git_has_remote

Has git remote

git_is_local_repository

Check if this folder is a local Git repository

git_tag_sync_check_create_push

sync, check, create, push git tag

github_api_create_new_release

Create new release on Github

github_api_upload_asset_to_release

Upload asset to github release

hash_text

Hash text

home_dir

home_dir() using the home crate.

init_repository_if_needed

Init repository if needed

now_utc_date_iso

UTC date in iso standard like 2024-12-31

publish_to_crates_io_with_secret_token

Publish to crates.io

run_shell_command

Run one shell command

run_shell_command_output

run one shell command and return ShellOutput {exit_status, stdout, stderr}

run_shell_command_success

run one shell command and return true if success

traverse_dir_with_exclude_dir

Traverse dir and its sub-dir, but avoid excluded dirs.