Module cargo_auto_lib::utils_mod

various utilities

Functions

concatenate_vec_to_string

the original concat() function does not have a delimiter

file_exists

check if file exists

file_path_home_expand

expands the ~ for home_dir and returns expanded path

find_from

find from_pos

find_pos_end_data_before_delimiter

return the position of end of the delimited data before the delimiter

find_pos_start_data_after_delimiter

return the position of start of the delimited data after the delimiter

to_path

just a shorter way to get a Path from a String

traverse_dir_with_exclude_dir

traverse_dir_with_exclude_dir